requirements.txt
pyyaml>=6.0
pytest>=7.0
pytest-asyncio>=0.21.0 